+++ README.md
@@ -1,43 +1,14 @@
-[![bitHound Score](https://www.bithound.io/github/Granze/react-starterify/badges/score.svg)](https://www.bithound.io/github/zooniverse/zoo-react-starterify/master)
+# Zooniverse Map Explorer
 
-# Zooniverse React Starterify
+Let's build a map feature using React and Leaflet! This project shows you how.
 
-A minimal React JS application starter kit, based on [React Starterify](https://github.com/Granze/react-starterify).
-If you’re looking for a Redux starter kit, check [this out](https://github.com/zooniverse/zoo-reduxify)
-## Usage
+## Development
 
-__Install the dependencies:__
+* Target audience: Zooniverse devs (and other React devs) who want to add mapping functionality to their React projects.
+* This is built on top of the [Zooniverse React Starterify](https://github.com/zooniverse/zoo-react-starterify) project.
 
-`npm install`
+Starting the project:
 
-__Test:__
-
-```npm run test```
-
-__Development mode with livereload:__
-
-```npm run start```
-
-__When you are done, create a production-ready version of the JS bundle:__
-
-```npm run build```
-
-## Credits
-
-Thanks to [Granze](https://github.com/Granze) for the original [React Starterify](https://github.com/Granze/react-starterify), which is used under the [MIT License](http://opensource.org/licenses/MIT).
-
-## License
-
-Copyright 2015 Zooniverse
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-   http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
+1. `npm install` to install dependencies.
+2. `npm start` to start the project.
+3. `http://localhost:3000` to view the project on your web browser.
