+++ src/Index.js
@@ -2,7 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Router, Route } from 'react-router';
 import App from './components/App';
-import Map1 from './pages/Map-1';
+import Map1 from './pages/Map-1/Index';
+import Map1A from './pages/Map-1A/Index';
 
 // Todo: let's find a better way to include Styles,
 // currently Styles looks like an unused var to eslint
@@ -15,6 +16,7 @@ ReactDOM.render(
   <Router>
     <Route path="/" component={App}>
       <Route path="/map-1" component={Map1}/>
+      <Route path="/map-1a" component={Map1A}/>
     </Route>
   </Router>
   , document.getElementById('root')
